+++ public/Import-AdfFromFolder.ps1
@@ -37,7 +37,9 @@ function Import-AdfFromFolder {
 
     if ( !(Test-Path -Path $RootFolder) ) { Write-Error "ADFT0019: Folder '$RootFolder' doesn't exist." }
     
-    $adf.Location = $RootFolder
+    $adf.Location = Resolve-Path $RootFolder
+    $adf.PublishOptions = New-AdfPublishOption
+    Write-Verbose "Resolved RootFolder = $($adf.Location)"
 
     Import-AdfObjects -Adf $adf -All $adf.IntegrationRuntimes -RootFolder $RootFolder -SubFolder "integrationRuntime" | Out-Null
     Write-Host ("IntegrationRuntimes: {0} object(s) loaded." -f $adf.IntegrationRuntimes.Count)