+++ test/PublishOption.Tests.ps1
@@ -63,9 +63,57 @@ InModuleScope azure.datafactory.tools {
                 $script:opt.Excludes.ContainsKey('*.SharedIR*') | Should -Be $true
                 $script:opt.Excludes.ContainsKey('*.LS_SqlServer_DEV19_AW2017') | Should -Be $true
             }
-
         }
 
+        Context 'When called with empty file in FilterFilePath' {
+            It 'Should not throw exception' {
+                {
+                    $script:opt = New-AdfPublishOption -FilterFilePath "$SrcFolder\deployment\filter.empty.txt"
+                } | Should -Not -Throw 
+            }
+            It 'Should add no rules to Includes and Excludes' {
+                $script:opt.Includes.Count | Should -Be 0
+                $script:opt.Excludes.Count | Should -Be 0
+            }
+        }
         
     } 
+
+    Describe 'New-AdfPublishOption' -Tag 'Unit' {
+
+        Context 'When new object created' {
+            It 'Should "TriggerStopMethod" = AllEnabled be default' {
+                $script:result = New-AdfPublishOption
+                $script:result.TriggerStopMethod | Should -Be 'AllEnabled'
+            }
+            It '"TriggerStopMethod" should be changeable' {
+                $script:result.TriggerStopMethod = [TriggerStopTypes]::DeployableOnly
+                $script:result.TriggerStopMethod | Should -Be 'DeployableOnly'
+                $script:result.TriggerStopMethod = 'AllEnabled'
+                $script:result.TriggerStopMethod | Should -Be 'AllEnabled'
+            }
+            It '"TriggerStopMethod" should not accept invalid values' {
+                { $script:result.TriggerStopMethod = [TriggerStopTypes]::SomethingInvalid } | Should -Throw 
+                { $script:result.TriggerStopMethod = 'SomethingInvalid' } | Should -Throw 
+            }
+        }
+
+        Context 'When new object created' {
+            It 'Should "TriggerStartMethod" = BasedOnSourceCode be default' {
+                $script:result = New-AdfPublishOption
+                $script:result.TriggerStartMethod | Should -Be 'BasedOnSourceCode'
+            }
+            It '"TriggerStartMethod" should be changeable' {
+                $script:result.TriggerStartMethod = [TriggerStartTypes]::KeepPreviousState
+                $script:result.TriggerStartMethod | Should -Be 'KeepPreviousState'
+                $script:result.TriggerStartMethod = 'BasedOnSourceCode'
+                $script:result.TriggerStartMethod | Should -Be 'BasedOnSourceCode'
+            }
+            It '"TriggerStartMethod" should not accept invalid values' {
+                { $script:result.TriggerStartMethod = [TriggerStartTypes]::SomethingInvalid } | Should -Throw 
+                { $script:result.TriggerStartMethod = 'SomethingInvalid' } | Should -Throw 
+            }
+        }
+
+    }
 }