+++ changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.6.2] - 2023-07-17
+### Added
+* Stop and restart only changed triggers #264 #332
+* New option to Stop/Start only the triggers that are already Started #291
+### Fixed
+* `RootFolder` must be absolute otherwise temp files cannot be written #335
+* Catch InvalidOperation exception when reading empty FilterFilePath in New-AdfPublishOption.ps1 #338
+* Purview configuration gets overwritten when deploy Global Parameters using ADF deployment task #343
+* Incremental Deployment seems to not work when factory folder does not exist #346
+* The publish cmdlet tries to delete enabled (active) trigger (not in source) when TriggerStartMethod = KeepPreviousState #350
+
 ## [1.5.0] - 2023-06-18
 ### Added
 * Remove pipeline's from check (`Test-AdfCode`) for dashes in name #309